+++ src/main/java/exception/httpVersionNotSupported/HttpVersionNotSupported.java
@@ -1,7 +1,7 @@
 package exception.httpVersionNotSupported;
 
 import exception.CustomException;
-import webserver.Constants.HttpStatus;
+import webserver.http.Constants.HttpStatus;
 
 public abstract class HttpVersionNotSupported extends CustomException {
     protected static final HttpStatus httpStatus = HttpStatus.HTTP_VERSION_NOT_SUPPORTED;
