+++ src/main/java/exception/internalServerError/InternalServerException.java
@@ -1,7 +1,7 @@
 package exception.internalServerError;
 
 import exception.CustomException;
-import webserver.Constants.HttpStatus;
+import webserver.http.Constants.HttpStatus;
 
 public abstract class InternalServerException extends CustomException {
     protected static final HttpStatus httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
