+++ src/main/java/webserver/http/Constants/ContentType.java
@@ -1,4 +1,4 @@
-package webserver.Constants;
+package webserver.http.Constants;
 
 public enum ContentType {
