+++ src/main/java/webserver/http/Constants/HeaderField.java
@@ -1,4 +1,4 @@
-package webserver.Constants;
+package webserver.http.Constants;
 
 public class HeaderField {
