+++ src/main/java/webserver/http/Constants/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver.Constants;
+package webserver.http.Constants;
 
 public enum HttpMethod {
