+++ src/main/java/webserver/http/Constants/HttpStatus.java
@@ -1,4 +1,4 @@
-package webserver.Constants;
+package webserver.http.Constants;
 
 import static support.utils.StringUtils.*;
