+++ src/main/java/webserver/http/Constants/HttpVersion.java
@@ -1,4 +1,4 @@
-package webserver.Constants;
+package webserver.http.Constants;
 
 import exception.httpVersionNotSupported.InvalidVersionException;
