+++ src/main/java/webserver/http/Header.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import java.util.Arrays;
 import java.util.HashMap;
