+++ src/main/java/webserver/view/view/View.java
@@ -1,8 +1,8 @@
 package webserver.view.view;
 
 import exception.internalServerError.FileRenderException;
-import webserver.request.HttpRequest;
-import webserver.response.HttpResponse;
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
 
 import java.io.DataOutputStream;
 import java.util.Map;
